[fixed] buffs having chance to set players max HAM negative upon removal - Mantis #8790 [fixed] channel force tisk task not accounting for buff removal - Manthis #8790

+++ MMOCoreORB/src/server/zone/objects/creature/CreatureObjectImplementation.cpp
@@ -1064,8 +1064,7 @@ int CreatureObjectImplementation::inflictDamage(TangibleObject* attacker, int da
 
 
 	int currentValue = hamList.get(damageType);
-
-	int newValue = currentValue - (int) damage;
+	int newValue = currentValue - damage;
 
 	if (!destroy && newValue <= 0)
 		newValue = 1;
